requirements.txt
Faker==18.11.1
install==1.3.5
numpy==1.25.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
